bdf  - cleanup to break_word_by_trailing_integer  - fixing spoint.points bug  - adding idtype to get_pid_to_node_ids_and_elements_array  - starting on supporting None nodes in get_pid_to_node_ids_and_elements_array

+++ pyNastran/bdf/test/test_bdf.py
@@ -557,7 +557,7 @@ def run_fem1(fem1, bdf_model, out_model, mesh_form, xref, punch, sum_load, size,
                                consider_0d=True, consider_0d_rigid=True,
                                consider_1d=True, consider_2d=True, consider_3d=True)
                 fem1.get_dependent_nid_to_components()
-                fem1.get_pid_to_node_ids_and_elements_array(pids=None, etypes=None)
+                fem1.get_pid_to_node_ids_and_elements_array(pids=None, etypes=None, idtype='int32')
                 fem1.get_property_id_to_element_ids_map()
                 fem1.get_material_id_to_property_ids_map()
                 fem1.get_element_ids_list_with_pids(pids=None)
